fetched wallet balance for verification and dispatcher
