+++ js/web/pages/admin.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import OnboardAdmin from '../components/OnboardAdmin'
+import Admin from '../components/Admin'
 import Box from '@mui/material/Box'
 const GetStartedAdminPage = () => {
   return (
     <Box>
-      <OnboardAdmin />
+      <Admin />
     </Box>
   )
 }
